updated comments following python guidelines

+++ app/lib/db_actions.py
@@ -77,15 +77,16 @@ def set_user_active(token):
     try:
         # Cerco il token ed abilito l'utente
         user = db.session.query(User).filter(Token.token == token, User.email == Token.email).first()
-        user.is_active = True
-        db.session.flush()
-        # elimino il token non più necessario
-        if user.is_active:
-            Token.query.filter(Token.token == token).delete()
+        if user:
+            user.is_active = True
+            db.session.flush()
+            # elimino il token non più necessario
+            if user.is_active:
+                Token.query.filter(Token.token == token).delete()
 
-        db.session.commit()
+            db.session.commit()
 
-        return user.is_active
+            return user.is_active
     except exc.SQLAlchemyError as e:
         print(type(e))
         db.session.rollback()
@@ -293,3 +294,12 @@ def get_building_from_lesson(id_course):
     building = Building.query.filter(Building.id_building == classroom.id_building).first()
 
     return building
+
+def get_users_from_course(id_course):
+    """
+    Ritorna l'unione di tutti gli id_user di professori e studenti presenti in un corso sotto forma di lista
+    """
+    a = db.session.query(StudentsCourses.id_student).filter(StudentsCourses.id_course == id_course)
+    b = db.session.query(TeachersCourses.id_teacher).filter(TeachersCourses.id_course == id_course)
+    res = a.union(b).all()
+    return [value for value, in res]
